% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuitscaper-package.R
\docType{package}
\name{circuitscaper-package}
\alias{circuitscaper}
\alias{circuitscaper-package}
\title{circuitscaper: 'Circuitscape' and 'Omniscape' Connectivity Analysis via 'Julia'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an R-native interface to the 'Circuitscape.jl' and 'Omniscape.jl' 'Julia' packages for landscape connectivity modeling using circuit theory. Users work entirely in R with familiar objects (SpatRaster, file paths) while 'Julia' handles computation invisibly. Supports all four 'Circuitscape' modes (pairwise, one-to-all, all-to-one, advanced) and 'Omniscape' moving-window analysis. Methods are described in McRae (2006) \doi{10.1111/j.0014-3820.2006.tb00500.x} and Landau et al. (2021) \doi{10.21105/joss.02829}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewkling/circuitscaper}
  \item \url{https://matthewkling.github.io/circuitscaper/}
  \item Report bugs at \url{https://github.com/matthewkling/circuitscaper/issues}
}

}
\author{
\strong{Maintainer}: Matthew Kling \email{mattkling@berkeley.edu} [copyright holder]

Authors:
\itemize{
  \item Matthew Kling \email{mattkling@berkeley.edu} [copyright holder]
}

}
\keyword{internal}
